requirements.txt
aiogram==3.20.0.post0
beautifulsoup4==4.13.4
matplotlib==3.7.2
pathvalidate==3.2.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
webdriver-manager==4.0.2
pandas==1.5.3

